checking if correction terms scale wrong with m_op

+++ src/markovanalyzer/polyspectra_calculator.py
@@ -845,7 +845,7 @@ class System:  # (SpectrumCalculator):
         self.freq[order] = f_data
 
         n_states = self.transtion_matrix.shape[0]
-        self.eigvals, self.eigvecs = eig(self.transtion_matrix)
+        self.eigvals, self.eigvecs = eig(self.transtion_matrix.astype(dtype=np.complex128))
         self.eigvecs_inv = inv(self.eigvecs)
 
         self.eigvals = self.eigvals.astype(dtype=np.complex128)
